refactor: item 조회, 메인 item 조회 리팩토링

+++ server/module-core/src/main/java/com/team33/modulecore/cart/application/CartService.java
@@ -57,7 +57,7 @@ public class CartService {
         for (ItemCart itemCart : itemCarts) {
             int quantity = itemCart.getItem().getPrice();
             int price = itemCart.getQuantity();
-            int discountRate = itemCart.getItem().getDiscountRate();
+            double discountRate = itemCart.getItem().getDiscountRate();
 
             totalDiscountPrice += (quantity * price * discountRate / 100);
         }
